requirements_v2.txt
libigl
scipy
numpy
trimesh[easy]